+++ CHANGELOG.rst
@@ -36,6 +36,9 @@ version 0.5.1 [2017-12-01]
 
 - Add an option to control the audio bell volume on X11 systems
 
+- Add a command line switch to set the name part of the WM_CLASS window
+  property independently.
+
 
 version 0.5.0 [2017-11-19]
 ---------------------------